refactor(react-front-end): Break out SearchPage's reducer
Following a pattern I started in
openequella/react-kaltura-simpleuploader , this pushes the reducer and
supporting types out into a standalone file to assist with managing the
size of SearchPage.tsx as well as aiding any additional testing etc we
may wish to do.
