+++ react-front-end/tsrc/search/SearchPage.tsx
@@ -101,8 +101,10 @@ import {
   generateSearchPageOptionsFromQueryString,
   getPartialSearchOptions,
   getRawModeFromStorage,
+  SearchPageOptions,
   writeRawModeToStorage,
 } from "./SearchPageHelper";
+import { reducer, SearchPageSearchResult } from "./SearchPageReducer";
 
 // destructure strings import
 const { searchpage: searchStrings } = languageStrings;
@@ -110,20 +112,6 @@ const { title: dateModifiedSelectorTitle, quickOptionDropdown } =
   searchStrings.lastModifiedDateSelector;
 const { title: collectionSelectorTitle } = searchStrings.collectionSelector;
 const { title: displayModeSelectorTitle } = searchStrings.displayModeSelector;
-
-/**
- * Type of search options that are specific to Search page presentation layer.
- */
-export interface SearchPageOptions extends SearchOptions {
-  /**
-   * Whether to enable Quick mode (true) or to use custom date pickers (false).
-   */
-  dateRangeQuickModeEnabled: boolean;
-  /**
-   * How to display the search results - also determines the type of results.
-   */
-  displayMode: DisplayMode;
-}
 
 /**
  * Structure of data stored in browser history state, to capture the current state of SearchPage
@@ -138,76 +126,6 @@ interface SearchPageHistoryState {
    */
   filterExpansion: boolean;
 }
-
-/**
- * The types of SearchResultItem that we support within an `OEQ.Search.SearchResult`.
- */
-type SearchPageSearchResult =
-  | {
-      from: "item-search";
-      content: OEQ.Search.SearchResult<OEQ.Search.SearchResultItem>;
-    }
-  | {
-      from: "gallery-search";
-      content: OEQ.Search.SearchResult<GallerySearchResultItem>;
-    };
-
-type Action =
-  | { type: "init" }
-  | { type: "search"; options: SearchPageOptions; scrollToTop: boolean }
-  | {
-      type: "search-complete";
-      result: SearchPageSearchResult;
-      classifications: Classification[];
-    }
-  | { type: "error"; cause: Error };
-
-type State =
-  | { status: "initialising" }
-  | {
-      status: "searching";
-      options: SearchPageOptions;
-      previousResult?: SearchPageSearchResult;
-      previousClassifications?: Classification[];
-      scrollToTop: boolean;
-    }
-  | {
-      status: "success";
-      result: SearchPageSearchResult;
-      classifications: Classification[];
-    }
-  | { status: "failure"; cause: Error };
-
-const reducer = (state: State, action: Action): State => {
-  switch (action.type) {
-    case "init":
-      return { status: "initialising" };
-    case "search":
-      const prevResults =
-        state.status === "success"
-          ? {
-              previousResult: state.result,
-              previousClassifications: state.classifications,
-            }
-          : {};
-      return {
-        status: "searching",
-        options: action.options,
-        scrollToTop: action.scrollToTop,
-        ...prevResults,
-      };
-    case "search-complete":
-      return {
-        status: "success",
-        result: action.result,
-        classifications: action.classifications,
-      };
-    case "error":
-      return { status: "failure", cause: action.cause };
-    default:
-      throw new TypeError("Unexpected action passed to reducer!");
-  }
-};
 
 const SearchPage = ({ updateTemplate }: TemplateUpdateProps) => {
   const history = useHistory();
